commit 2

+++ src/Service/UserSoapService.php
@@ -10,34 +10,69 @@ use Symfony\Component\HttpFoundation\Request;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
-
+use Doctrine\Common\Collections\Collection;
 class UserSoapService extends AbstractController
 {
 
 	private $em;
-   
+    private $verifierJetonService;
     private $encoder;
   
-    public function __construct(EntityManagerInterface $em, UserPasswordEncoderInterface $encoder){
+    public function __construct(EntityManagerInterface $em, UserPasswordEncoderInterface $encoder, VerifierJetonService $verifierJetonService){
         $this->em = $em;
-        
+        $this->verifierJetonService = $verifierJetonService;
         $this->encoder = $encoder;
 
     }
 
-       public function list(){
+    /**
+    * @param string $emailUser 
+    * @return array
+    */
+    
+    public function list(string $emailUser){
+
+        if(!$this->verifierJetonService->verifierJeton($emailUser, "READ")){
+            return array('note' => "vous n'avez pas acces à ce service");
+        }
       
         $users = $this->em->getRepository(User::class)->findAll();
        
-        return $users;
+        $usersArray = array();
+        foreach ($users as  $user) {
+            $userArray = array();
+            $userArray[] = $user->getId();
+            $userArray[] = $user->getPrenom();
+            $userArray[] = $user->getNom();
+            $userArray[] = $user->getEmail();
+            $userArray[] = $user->getRoles();
+            $usersArray[] = $userArray;
+        }
+
+        return $usersArray;
     }
 
-    public function test($a, $b){
+    /**
+    * @param int $a
+    * @param int $b
+    * @return int
+    */
+    public function tr($a, $b){
         return $a+$b;
     }
-    public function create($prenom, $nom, $email){
-        var_dump("laye");
-        die();
+
+    /**
+    * @param string $prenom
+    * @param string $nom
+    * @param string $email
+    * @param string $emailUser 
+    * @return string 
+    */
+    public function create($prenom, $nom, $email, $emailUser){
+
+        if(!$this->verifierJetonService->verifierJeton($emailUser, "CREATE")){
+            return "vous n'avez pas acces à ce service";
+        }        
         $user = new User;
         $user->setPrenom($prenom);
         $user->setNom($nom);
@@ -53,12 +88,22 @@ class UserSoapService extends AbstractController
             $this->em->persist($user);
             $this->em->flush();
     
-            return "ok";
+            return "create";
 
         
        
     }
-    public function edit($id, $status){
+     /**
+    * @param int $id
+    * @param int $status
+    * @param string $emailUser
+    * @return string 
+    */
+    public function edit($id, $status, $emailUser){
+
+        if(!$this->verifierJetonService->verifierJeton($emailUser, "UPD")){
+            return "vous n'avez pas acces à ce service";
+        }
        $user = $this->em->getRepository(User::class)->find($id);
         if($status == 1){
             $user->setStatus(1);
@@ -71,10 +116,8 @@ class UserSoapService extends AbstractController
             $user->setRoles(["ROLE_USER"]);
         }
 
-        $this->em->flush();
-        $this->addFlash('success', 'Privilièeg Modifié');
+       
         
-        // return $this->redirectToRoute('task_success');
          return "ok";
            
        
@@ -83,15 +126,51 @@ class UserSoapService extends AbstractController
         
      }
 
+    /**
+    * @param int $id
+    * @param string $emailUser 
+    * @return string 
+    */
    
-    public function delete($id){
-        dd("laye");
+    public function delete($id, $emailUser){
+
+        if(!$this->verifierJetonService->verifierJeton($emailUser, "DEL")){
+            return "vous n'avez pas acces à ce service";
+        }
         $user = $this->em->getRepository(User::class)->find($id);
         $this->em->remove($user);
         $this->em->flush();
            
         
-        return $id;
+        return "ok";
+    }
+
+    /**
+    * @param string $email
+    * @param string $passe
+    * @return string
+    */
+
+    public function authentificate($email, $passe){
+
+        $user = $this->em->getRepository(User::class)->findOneBy(['email' => $email]);
+
+        $passe = $this->encoder->encodePassword(
+                $user,
+                $passe
+            );
+
+        $users = $this->em->getRepository(User::class)->findAll();
+
+        foreach ($users as $user) {
+
+            if($user->getPassword() == $passe){
+                return "oui";
+            }
+        }
+
+        return "non";
+
     }
 
 }
